cast a time_t so that it does not conflict with the print format.
This change was already done on trunk so this change needs to be
blocked from merging.



git-svn-id: https://origsvn.digium.com/svn/asterisk/branches/1.4@72453 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ apps/app_meetme.c
@@ -2258,7 +2258,7 @@ bailoutandtrynormal:
 				      user->user_no,
 				      S_OR(user->chan->cid.cid_num, "<unknown>"),
 				      S_OR(user->chan->cid.cid_name, "<unknown>"),
-				      (now - user->jointime));
+				      (long)(now - user->jointime));
 		}
 
 		conf->users--;
